+++ aha/util/regress.py
@@ -93,7 +93,7 @@ def test_sparse_app(testname, width, height, test=""):
     print(f"--- {test} - glb testing")
     start = time.time()
     buildkite_call(
-        ["aha", "glb", app_path, "--sparse", "--sparse-test-name", testname], env=env_vars,
+        ["aha", "test", app_path, "--sparse", "--sparse-test-name", testname], env=env_vars,
     )
     time_test = time.time() - start
 
@@ -116,7 +116,7 @@ def test_dense_app(test, width, height, layer=None, env_parameters=""):
         pass
 
     start = time.time()
-    buildkite_call(["aha", "halide", test, "--env-parameters", env_parameters] + layer_array)
+    buildkite_call(["aha", "map", test, "--env-parameters", env_parameters] + layer_array)
     time_compile = time.time() - start
 
     print(f"--- {test} - mapping")
@@ -125,7 +125,7 @@ def test_dense_app(test, width, height, layer=None, env_parameters=""):
     buildkite_call(
         [
             "aha",
-            "pipeline",
+            "pnr",
             test,
             "--width", str(width),
             "--height", str(height),
@@ -136,7 +136,7 @@ def test_dense_app(test, width, height, layer=None, env_parameters=""):
 
     print(f"--- {test} - glb testing")
     start = time.time()
-    buildkite_call(["aha", "glb", test])
+    buildkite_call(["aha", "test", test])
     time_test = time.time() - start
 
     return time_compile, time_map, time_test
